+++ frontend/src/components/Agreements/ProjectOfficerSelect/ProjectOfficerSelect.jsx
@@ -1,15 +1,13 @@
-import { useState } from "react";
-import PropTypes from "prop-types";
-import { useGetUsersQuery } from "../../api/opsAPI";
+import { useEffect, useState } from "react";
+import { useGetUsersQuery } from "../../../api/opsAPI";
 
-export const TeamMemberSelect = ({
-    className,
-    selectedProjectOfficer,
-    selectedTeamMembers,
-    setSelectedTeamMembers,
-}) => {
+export const ProjectOfficerSelect = ({ selectedProjectOfficer, setSelectedProjectOfficer }) => {
     const { data: users, error: errorUsers, isLoading: isLoadingUsers } = useGetUsersQuery();
-    const [inputValue, setInputValue] = useState("");
+    const [inputValue, setInputValue] = useState(selectedProjectOfficer?.full_name ?? "");
+
+    useEffect(() => {
+        setInputValue(selectedProjectOfficer?.full_name ?? "");
+    }, [selectedProjectOfficer]);
 
     if (isLoadingUsers) {
         return <div>Loading...</div>;
@@ -18,50 +16,48 @@ export const TeamMemberSelect = ({
         return <div>Oops, an error occurred</div>;
     }
 
-    const remainingUsers = users.filter(
-        (user) =>
-            user.id !== selectedProjectOfficer?.id && // Check if the user is not a selected project officer
-            !selectedTeamMembers.some((teamMember) => teamMember.id === user.id) // Check if the user is not already a team member
-    );
-
     const onChangeSelect = (userId = 0) => {
         if (userId === 0) {
+            setSelectedProjectOfficer(null);
             return;
         }
-        const selected = users.find((user) => user.id === userId);
-        setSelectedTeamMembers((prevState) => {
-            return [...prevState, selected];
-        });
-        setInputValue("");
+        const selected = users[userId - 1];
+        setSelectedProjectOfficer({ ...selected });
     };
 
     return (
-        <div className={`usa-fieldset ${className}`}>
-            <label className="usa-label margin-top-0" htmlFor="team-member-select" id="team-member-select-label">
-                Team Members
+        <fieldset className="usa-fieldset">
+            <label
+                className="usa-label margin-top-0"
+                htmlFor="project-officer-select"
+                id="project-officer-select-label"
+            >
+                Project Officer
             </label>
             <div className="usa-combo-box width-card-lg" data-enhanced="true">
                 <select
                     className="usa-select usa-sr-only usa-combo-box__select "
-                    name="team-member-select"
+                    id="project-officer-select"
+                    name="project-officer-select"
                     aria-hidden="true"
                     tabIndex="-1"
-                    value={remainingUsers?.id}
+                    value={selectedProjectOfficer?.id}
                     onChange={(e) => onChangeSelect(Number(e.target.value))}
                     required
                 >
-                    {remainingUsers?.map((user) => (
+                    {users.map((user) => (
                         <option key={user?.id} value={user?.id}>
                             {user?.full_name || user?.email}
                         </option>
                     ))}
                 </select>
                 <input
-                    id="team-member-select"
-                    aria-owns="team-members--list"
-                    aria-controls="team-members--list"
+                    id="project-officer-select-input"
+                    aria-label="project-officer-select-input"
+                    aria-owns="users--list"
+                    aria-controls="users--list"
                     aria-autocomplete="list"
-                    aria-describedby="team-members--assistiveHint"
+                    aria-describedby="users--assistiveHint"
                     aria-expanded="false"
                     autoCapitalize="off"
                     autoComplete="off"
@@ -77,7 +73,9 @@ export const TeamMemberSelect = ({
                         type="button"
                         className="usa-combo-box__clear-input"
                         aria-label="Clear the select contents"
-                        onClick={() => onChangeSelect(0)}
+                        onClick={() => {
+                            setSelectedProjectOfficer(null);
+                        }}
                     >
                         &nbsp;
                     </button>
@@ -85,7 +83,7 @@ export const TeamMemberSelect = ({
                 <span className="usa-combo-box__input-button-separator">&nbsp;</span>
                 <span className="usa-combo-box__toggle-list__wrapper" tabIndex="-1">
                     <button
-                        id="team-member-select-toggle-list"
+                        id="project-officer-select-toggle-list"
                         type="button"
                         tabIndex="-1"
                         className="usa-combo-box__toggle-list"
@@ -97,20 +95,20 @@ export const TeamMemberSelect = ({
 
                 <ul
                     tabIndex="-1"
-                    id="team-members--list"
+                    id="users--list"
                     className="usa-combo-box__list"
                     role="listbox"
-                    aria-labelledby="team-member-select-label"
+                    aria-labelledby="project-officer-select-label"
                     hidden
                 >
                     {users?.map((user, index) => {
                         return (
                             <li
                                 key={user?.id}
-                                aria-setsize={remainingUsers?.length}
+                                aria-setsize={users?.length}
                                 aria-posinset={index + 1}
                                 aria-selected="false"
-                                id={`team-member-dynamic-select--list--option-${index}`}
+                                id={`project-officer-dynamic-select--list--option-${index}`}
                                 className="usa-combo-box__list-option"
                                 tabIndex={index === 0 ? "0" : "-1"}
                                 role="option"
@@ -123,17 +121,13 @@ export const TeamMemberSelect = ({
                 </ul>
 
                 <div className="usa-combo-box__status usa-sr-only" role="status"></div>
-                <span id="team-members--assistiveHint" className="usa-sr-only">
+                <span id="users--assistiveHint" className="usa-sr-only">
                     When autocomplete results are available use up and down arrows to review and enter to select. Touch
                     device users, explore by touch or with swipe gestures.
                 </span>
             </div>
-        </div>
+        </fieldset>
     );
 };
 
-export default TeamMemberSelect;
-
-TeamMemberSelect.propTypes = {
-    className: PropTypes.string,
-};
+export default ProjectOfficerSelect;
